refactor(suite-common): separate selectors logic
